Added device selecting list for SetDevices interface. Added arrow buttons for SetDevices interface.
git-svn-id: https://ccv-multi-cam.googlecode.com/svn@43 da66ed7f-4d6a-8cb4-a557-1474cfe16edc

+++ branches/test/CCV_Select_Camera/addons/ofxNCore/src/Controls/ofxGuiGrid.h
@@ -54,7 +54,7 @@ public:
 	void saveToXml();
 
 	int mWidthScale, mHeightScale;
-	int mGridWidth, mGridHeight;
+	float mGridWidth, mGridHeight;
 	int mXGrid, mYGrid, mBorder, mSpacing;
 
 	int mValue, mSelectedId;
